Module OpamParser

opam format parser

Raw OpamBaseParser entry points

val main : (Stdlib.Lexing.lexbuf -> OpamBaseParser.token) -> Stdlib.Lexing.lexbuf -> OpamParserTypes.file_name -> OpamParserTypes.opamfile

Principal parser: given a lexbuf and the filename it was read from, returns an OpamParserTypes.opamfile record parsed from it.

val value : (Stdlib.Lexing.lexbuf -> OpamBaseParser.token) -> Stdlib.Lexing.lexbuf -> OpamParserTypes.value

Lower-level function just returning a single OpamParserTypes.value from a given lexer.

val string : string -> OpamParserTypes.file_name -> OpamParserTypes.opamfile

File parsers

Parse the content of a file already read to a string. Note that for CRLF-detection to work on Windows, it is necessary to read the original file using binary mode on Windows!

val channel : Stdlib.in_channel -> OpamParserTypes.file_name -> OpamParserTypes.opamfile

Parse the content of a file from an already-opened channel. Note that for CRLF-detection to work on Windows, it is necessary for the channel to be in binary mode!

val file : OpamParserTypes.file_name -> OpamParserTypes.opamfile

Parse the content of a file. The file is opened in binary mode, so CRLF-detection works on all platforms.

value parsers

val value_from_string : string -> OpamParserTypes.file_name -> OpamParserTypes.value

Parse the first value in the given string. file_name is used for lexer positions.

val value_from_channel : Stdlib.in_channel -> OpamParserTypes.file_name -> OpamParserTypes.value

Parse the first value from the given channel. file_name is used for lexer positions.

val value_from_file : OpamParserTypes.file_name -> OpamParserTypes.value

Parse the first value from the given file.