Module OpamPrinter.FullPos

OpamPrinter transitional module with full position types

Printers for the value and opamfile formats

val relop_kind : [< OpamParserTypes.FullPos.relop_kind ] -> string

Converts OpamParserTypes.FullPos.relop_kind to its string representation (=, !=, ..., ~).

val logop_kind : [< OpamParserTypes.FullPos.logop_kind ] -> string

Converts OpamParserTypes.FullPos.logop_kind to its string representation (& and |).

val pfxop_kind : [< OpamParserTypes.FullPos.pfxop_kind ] -> string

Converts OpamParserTypes.FullPos.logop_kind to its string representation (& and |).

val env_update_op_kind : OpamParserTypes.FullPos.env_update_op_kind -> string

Converts OpamParserTypes.FullPos.env_update_op_kind to its string representation (=, +=, ..., =:).

val relop : OpamParserTypes.FullPos.relop -> string

Converts OpamParserTypes.FullPos.relop to its string representation (=, !=, ..., ~).

val logop : OpamParserTypes.FullPos.logop -> string

Converts OpamParserTypes.FullPos.logop to its string representation (& and |).

val pfxop : OpamParserTypes.FullPos.pfxop -> string

Converts OpamParserTypes.FullPos.pfxop to its string representation (! and ?).

val env_update_op : OpamParserTypes.FullPos.env_update_op -> string

Converts OpamParserTypes.FullPos.env_update_op to its string representation (=, +=, ..., =:).

val value : OpamParserTypes.FullPos.value -> string

Converts value to a string always using LF-encoding of newlines.

val value_list : OpamParserTypes.FullPos.value list OpamParserTypes.FullPos.with_pos -> string

Converts a list of values to a string always using LF-encoding of newlines.

val items : OpamParserTypes.FullPos.opamfile_item list -> string

Converts a list of opam field/sections to a string.

raises Invalid_argument
if "opam-version" is greater than "2.0" and not solely the first item.

val opamfile : OpamParserTypes.FullPos.opamfile -> string

Converts an opamfile to a string.

raises Invalid_argument
if "opam-version" is greater than "2.0" and not solely the first item.

val format_opamfile : Format.formatter -> OpamParserTypes.FullPos.opamfile -> unit

Writes an opamfile to a Format.formatter. The function ensures that all newlines are sent using Format's break instructions (and so ultimately are processed with the out_newline function of the formatter) but it is the responsibility of the caller to ensure that the formatter is configured for the required output, if necessary.

raises Invalid_argument
if "opam-version" is greater than "2.0" and not solely the first item.

Normalised output for opam syntax files

module Normalise : sig ... end

opam normalised file format, for signatures.

Format-preserving reprinter

module Preserved : sig ... end

Random utility functions

val value_equals : OpamParserTypes.FullPos.value -> OpamParserTypes.FullPos.value -> bool

Compares structurally, without considering file positions

val opamfile_item_equals : OpamParserTypes.FullPos.opamfile_item -> OpamParserTypes.FullPos.opamfile_item -> bool

Compares structurally, without considering file positions