Module FullPos.Preserved

val items : string -> OpamParserTypes.FullPos.opamfile_item list -> OpamParserTypes.FullPos.opamfile_item list -> string

items str orig_its its converts its to string, while attempting to preserve the layout and comments of the original str for unmodified elements. The function assumes that str parses to the items orig_its.

val opamfile : ?format_from:OpamParserTypes.FullPos.file_name -> OpamParserTypes.FullPos.opamfile -> string

opamfile f converts f to string, respecting the layout and comments in the corresponding on-disk file for unmodified items. format_from can be specified instead of using the filename specified in f.